backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic
pydantic[email]
pydantic-settings
python-dotenv==1.0.1
python-multipart==0.0.9
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
sqlalchemy_utils==0.41.2
langchain==0.3.25
langchain-openai==0.3.19
langchain-community==0.3.25
langchain-core==0.3.65
langchain-text-splitters==0.3.8
PyJWT==2.8.0
openai==1.68.2
python-jose[cryptography]==3.3.0
bcrypt==4.3.0
asyncpg==0.29.0
greenlet==3.2.3
langgraph==0.5.1
redis==5.0.1
langgraph-checkpoint-redis==0.0.8
fastmcp

# Email RAG Pipeline
chromadb
sentence-transformers
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
beautifulsoup4

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
